% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{j1}
\alias{j1}
\title{J1 Data}
\format{An object of class \code{data.frame} with 528 rows and 4 columns.}
\source{
\url{https://sites.google.com/site/kanolabweb/home/textbook}
}
\usage{
j1
}
\description{
A dataset containing transfer data of J1 league players, 2011-2012
}
\keyword{datasets}
